Add AminoAcid calibrator
Still need to add a function that can override the settings inside of the MonteCarloTitration class to use the calibrated energies. Also need to add convergence checks/run till converged options.
