+++ constph/calibration-systems/asp-explicit.cpin
@@ -1,0 +1,17 @@
+&CNSTPH
+ CHRGDAT=-0.4157,0.2719,0.0341,0.0864,-0.1783,-0.0122,-0.0122,0.7994,-0.8014,
+ -0.8014,0.0,0.5973,-0.5679,0.0,0.0,0.0,-0.4157,0.2719,0.0341,0.0864,-0.0316,
+ 0.0488,0.0488,0.6462,-0.5554,-0.6376,0.4747,0.5973,-0.5679,0.0,0.0,0.0,-0.4157,
+ 0.2719,0.0341,0.0864,-0.0316,0.0488,0.0488,0.6462,-0.5554,-0.6376,0.0,0.5973,
+ -0.5679,0.4747,0.0,0.0,-0.4157,0.2719,0.0341,0.0864,-0.0316,0.0488,0.0488,
+ 0.6462,-0.6376,-0.5554,0.0,0.5973,-0.5679,0.0,0.4747,0.0,-0.4157,0.2719,0.0341,
+ 0.0864,-0.0316,0.0488,0.0488,0.6462,-0.6376,-0.5554,0.0,0.5973,-0.5679,0.0,0.0,
+ 0.4747,
+ PROTCNT=0,1,1,1,1,
+ RESNAME='System: Unknown','Residue: AS4 2',
+ RESSTATE=0,
+ STATEINF(0)%FIRST_ATOM=7, STATEINF(0)%FIRST_CHARGE=0, 
+ STATEINF(0)%FIRST_STATE=0, STATEINF(0)%NUM_ATOMS=16, STATEINF(0)%NUM_STATES=5, 
+ STATENE=0,38.7598760021,38.7598760021,38.7598760021,38.7598760021,
+ TRESCNT=1,CPHFIRST_SOL=29, CPH_IGB=2, CPH_INTDIEL=1.0, 
+/
