+++ constph/calibration-systems/asp-implicit.state.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".189278057868595" y=".12074797103063069" z=".06160079929887189"/>
+		<Position x=".1967018851540583" y=".2185175218044065" z=".014063739694148175"/>
+		<Position x=".1278890108868892" y=".2879409064429012" z=".06221419119152595"/>
+		<Position x=".17154971604570513" y=".20981016821385165" z="-.09160121219570554"/>
+		<Position x=".33767577197883203" y=".27088424937472616" z=".027453270599821587"/>
+		<Position x=".42329697830181134" y=".20463170321604388" z=".08576480068012633"/>
+		<Position x=".3613933293889335" y=".38999018913892425" z="-.026769747497631932"/>
+		<Position x=".2842168351849951" y=".4374361441157176" z="-.07144293022945895"/>
+		<Position x=".4906613055274841" y=".4598926413863216" z="-.024616362901358158"/>
+		<Position x=".5467379265123293" y=".4280460159994415" z=".06330910326126062"/>
+		<Position x=".5708695457111492" y=".42061392024478234" z="-.14959053621443844"/>
+		<Position x=".5833786323195739" y=".31232276058657715" z="-.15162881647791582"/>
+		<Position x=".5148163021586064" y=".4497708365625687" z="-.2384649744770032"/>
+		<Position x=".7097119952071715" y=".486087618722011" z="-.1522621279583084"/>
+		<Position x=".7941057050524613" y=".46295223320928663" z="-.046476117675416165"/>
+		<Position x=".7423945488223725" y=".5640040292036463" z="-.2604264242573922"/>
+		<Position x=".8305778153602923" y=".6013422730060849" z="-.2550496165048215"/>
+		<Position x=".47317220288345635" y=".6129910728937029" z="-.016223790655271183"/>
+		<Position x=".37650320631792705" y=".6687288053749708" z="-.0694045160171425"/>
+		<Position x=".6717680606820621" y=".5692559596053827" z="-.3251028027817445"/>
+		<Position x=".8788797497842942" y=".5069560126530547" z="-.05521902807374219"/>
+		<Position x=".7551681499131231" y=".4062704055311693" z=".020384549708630743"/>
+		<Position x=".5676194940888696" y=".681611696817347" z=".04938629278034622"/>
+		<Position x=".6448336699251058" y=".6295868936070281" z=".08875131425611646"/>
+		<Position x=".5682973763202144" y=".8268198960379872" z=".06620455860659522"/>
+		<Position x=".47757369154749413" y=".8589848800294384" z=".11744629273421973"/>
+		<Position x=".6548086390981117" y=".8572459720477489" z=".12519129458533151"/>
+		<Position x=".573043406195689" y=".8756598885810438" z="-.0311767858945043"/>
+	</Positions>
+</State>
