+++ constph/calibration-systems/cys-implicit.cpin
@@ -1,0 +1,12 @@
+&CNSTPH
+ CHRGDAT=-0.4157,0.2719,0.0213,0.1124,-0.1231,0.1112,0.1112,-0.3119,0.1933,
+ 0.5973,-0.5679,-0.4157,0.2719,0.0213,0.1124,-0.3593,0.1122,0.1122,-0.8844,0.0,
+ 0.5973,-0.5679,
+ PROTCNT=1,0,
+ RESNAME='System: Unknown','Residue: CYS 2',
+ RESSTATE=0,
+ STATEINF(0)%FIRST_ATOM=7, STATEINF(0)%FIRST_CHARGE=0, 
+ STATEINF(0)%FIRST_STATE=0, STATEINF(0)%NUM_ATOMS=11, STATEINF(0)%NUM_STATES=2, 
+ STATENE=89.1511443544,0,
+ TRESCNT=1,
+/
