+++ constph/calibration-systems/cys-implicit.state.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".19497880707327356" y=".12590275390602576" z=".07542227498507215"/>
+		<Position x=".2014742208783421" y=".22050345575195482" z=".021674692515940917"/>
+		<Position x=".13345853814016262" y=".29284814936515763" z=".06662336056526418"/>
+		<Position x=".17436081979409984" y=".20516895469431426" z="-.08277621339958485"/>
+		<Position x=".3427478373165202" y=".27358791537837607" z=".0291419690520408"/>
+		<Position x=".4295047542408177" y=".2102378161737571" z=".08918340460827814"/>
+		<Position x=".36535004009547545" y=".38973794483129687" z="-.031963835813905994"/>
+		<Position x=".2871745472701533" y=".4347603969271295" z="-.07738078576262943"/>
+		<Position x=".494275889752788" y=".45965143081244586" z="-.03578138596288341"/>
+		<Position x=".553020033275395" y=".4304344887719215" z=".051267437744178725"/>
+		<Position x=".5690486530621716" y=".4148140540965104" z="-.16240922582082626"/>
+		<Position x=".5669711209979553" y=".30600364219448317" z="-.16859462480700088"/>
+		<Position x=".5194414959247828" y=".45548062141873896" z="-.25054386270575085"/>
+		<Position x=".7422637289962157" y=".46948813474853246" z="-.15805461640423987"/>
+		<Position x=".7784673772321364" y=".4165243545282989" z="-.27525801231939734"/>
+		<Position x=".4753236404522559" y=".6128653489599987" z="-.03049455471372111"/>
+		<Position x=".37392473670716647" y=".6656393371254321" z="-.07774301588219173"/>
+		<Position x=".5728798597424384" y=".6849639522120998" z=".02633116558466975"/>
+		<Position x=".6540716721961342" y=".6351818745320477" z=".0599686900445204"/>
+		<Position x=".5718814109331707" y=".8305685345570194" z=".04012538697154253"/>
+		<Position x=".485363785388788" y=".8620469614792" z=".09848500452883457"/>
+		<Position x=".6625185079090976" y=".8639829743178096" z=".0906256305700903"/>
+		<Position x=".5670388404080455" y=".8773812079706707" z="-.05819628496028401"/>
+	</Positions>
+</State>
