+++ constph/calibration-systems/glu-implicit.state.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".19777967095375795" y=".12661753850838184" z=".0711915523934154"/>
+		<Position x=".20387035903475303" y=".22154250206822954" z=".01803321441574162"/>
+		<Position x=".13166304008147667" y=".291603854400861" z=".05987645571299076"/>
+		<Position x=".18223965272903" y=".20579915887208008" z="-.08759665639123446"/>
+		<Position x=".3428167265712403" y=".27881491784290924" z=".03226511739648931"/>
+		<Position x=".42901599797214046" y=".21794705937585193" z=".09538084269026996"/>
+		<Position x=".36426213116951117" y=".3961258858615876" z="-.026585838808082247"/>
+		<Position x=".28690198056956856" y=".4391201937635513" z="-.07527681646774359"/>
+		<Position x=".491660191133396" y=".46962476241464096" z="-.025017183636666093"/>
+		<Position x=".547484324428085" y=".44187978937992023" z=".06445866595937277"/>
+		<Position x=".5744217360663536" y=".4282893270470018" z="-.1481073281610009"/>
+		<Position x=".5822264032693149" y=".31956339499183584" z="-.1499576208798896"/>
+		<Position x=".5229640895530528" y=".4602398133842619" z="-.23874950528550373"/>
+		<Position x=".7163023716733699" y=".48597972983542137" z="-.148063643664539"/>
+		<Position x=".7117041965649339" y=".5946480000541846" z="-.15574225200912584"/>
+		<Position x=".7652284193734601" y=".4606529453469833" z="-.05394649904268571"/>
+		<Position x=".7974636099752762" y=".42979484888180397" z="-.2659146098512388"/>
+		<Position x=".7889109287985788" y=".4936892696909407" z="-.3873497579892752"/>
+		<Position x=".8755412897873665" y=".3187181588744197" z="-.24427174545210223"/>
+		<Position x=".9231836364063531" y=".29001754030793747" z="-.3224089012093229"/>
+		<Position x=".46852071639359766" y=".6221444127344375" z="-.020914454846547692"/>
+		<Position x=".37476763650226896" y=".6736425544755564" z="-.08295610655921509"/>
+		<Position x=".8704248482502287" y=".2867122054087785" z="-.15400517552210857"/>
+		<Position x=".842560265076627" y=".452855287545468" z="-.4555639470206559"/>
+		<Position x=".7303295269865442" y=".5696030149781632" z="-.3853608260849392"/>
+		<Position x=".5545901095108773" y=".6952442521605506" z=".051275290526210125"/>
+		<Position x=".6292137215092727" y=".6469211715374222" z=".09933571049146044"/>
+		<Position x=".5487291660535014" y=".8407292638539111" z=".06484896786156079"/>
+		<Position x=".45277196789825763" y=".8704982011155018" z=".10723756595663857"/>
+		<Position x=".6285794132818394" y=".8755079425240939" z=".13045275133146292"/>
+		<Position x=".5602874589685376" y=".8878145450232209" z="-.03282438290193299"/>
+	</Positions>
+</State>
