+++ constph/calibration-systems/his-implicit.cpin
@@ -1,0 +1,15 @@
+&CNSTPH
+ CHRGDAT=-0.3479,0.2747,-0.1354,0.1212,-0.0414,0.081,0.081,-0.0012,-0.1513,
+ 0.3866,-0.017,0.2681,-0.1718,0.3911,-0.1141,0.2317,0.7341,-0.5894,-0.3479,
+ 0.2747,-0.1354,0.1212,-0.111,0.0402,0.0402,-0.0266,-0.3811,0.3649,0.2057,
+ 0.1392,-0.5727,0.0,0.1292,0.1147,0.7341,-0.5894,-0.3479,0.2747,-0.1354,0.1212,
+ -0.1012,0.0367,0.0367,0.1868,-0.5432,0.0,0.1635,0.1435,-0.2795,0.3339,-0.2207,
+ 0.1862,0.7341,-0.5894,
+ PROTCNT=2,1,1,
+ RESNAME='System: Unknown','Residue: HIP 2',
+ RESSTATE=0,
+ STATEINF(0)%FIRST_ATOM=7, STATEINF(0)%FIRST_CHARGE=0, 
+ STATEINF(0)%FIRST_STATE=0, STATEINF(0)%NUM_ATOMS=18, STATEINF(0)%NUM_STATES=3, 
+ STATENE=0,-11.7770114534,-16.3478974337,
+ TRESCNT=1,
+/
