+++ constph/calibration-systems/his-implicit.state.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".1992018886095056" y=".1328651382483019" z=".11221653437693774"/>
+		<Position x=".2040393753195501" y=".22156534074138595" z=".04905585621708544"/>
+		<Position x=".13196239016445843" y=".29540926559463504" z=".08416077423076011"/>
+		<Position x=".1809799916892874" y=".19435807806980476" z="-.05394129856370777"/>
+		<Position x=".3430364779458945" y=".2807365068423635" z=".055097014126347335"/>
+		<Position x=".42940625146635825" y=".2281112983826851" z=".12535862154593752"/>
+		<Position x=".3641915183697395" y=".38993677101579266" z="-.018650834410468643"/>
+		<Position x=".2860369151525648" y=".42548908914582073" z="-.07184815128995174"/>
+		<Position x=".4911232122985805" y=".4643025024196547" z="-.026535719828381245"/>
+		<Position x=".5466747564503567" y=".4486088016930984" z=".0659299825884494"/>
+		<Position x=".5742520508636065" y=".40796122234169574" z="-.14334867386150074"/>
+		<Position x=".5842898602472073" y=".3001262043892107" z="-.13098396011788885"/>
+		<Position x=".521360782988744" y=".42547787671078896" z="-.23703831907772263"/>
+		<Position x=".7129186346740215" y=".4665701503814963" z="-.15305701165102817"/>
+		<Position x=".7575343456146408" y=".5562535021723628" z="-.2486457333942309"/>
+		<Position x=".7037343362079601" y=".5924086892241794" z="-.3261296790303951"/>
+		<Position x=".8863452206688305" y=".5855834629598905" z="-.22138583507167361"/>
+		<Position x=".949061802457617" y=".6524194934685946" z="-.2785575369651301"/>
+		<Position x=".9242358428797601" y=".5169662261605973" z="-.1114776984926798"/>
+		<Position x="1.0159393303533772" y=".5196240941738035" z="-.06916624142578982"/>
+		<Position x=".8166499735093082" y=".4417764275068172" z="-.0674118801590124"/>
+		<Position x=".8146637176284086" y=".3759017435947187" z=".018169602806321426"/>
+		<Position x=".46810334177164126" y=".6159139594549049" z="-.04076406073934704"/>
+		<Position x=".36538744077652635" y=".659140400039196" z="-.09414844230750333"/>
+		<Position x=".5645882020652845" y=".6966400009127267" z=".0048760126492753495"/>
+		<Position x=".6459661102992513" y=".6529834313626979" z=".04577964934469514"/>
+		<Position x=".5620805317951265" y=".8428408584464184" z="-.0020089595787279365"/>
+		<Position x=".5552800489039599" y=".8753546099886289" z="-.10582942963078416"/>
+		<Position x=".47612286459255787" y=".8814706570895013" z=".0527734239446097"/>
+		<Position x=".6530675088901725" y=".8838864445143643" z=".041791953473293834"/>
+	</Positions>
+</State>
