+++ constph/calibration-systems/lys-explicit.cpin
@@ -1,0 +1,14 @@
+&CNSTPH
+ CHRGDAT=-0.3479,0.2747,-0.24,0.1426,-0.0094,0.0362,0.0362,0.0187,0.0103,0.0103,
+ -0.0479,0.0621,0.0621,-0.0143,0.1135,0.1135,-0.3854,0.34,0.34,0.34,0.7341,
+ -0.5894,-0.3479,0.2747,-0.24,0.1426,-0.10961,0.034,0.034,0.06612,0.01041,
+ 0.01041,-0.03768,0.01155,0.01155,0.32604,-0.03358,-0.03358,-1.03581,0.0,
+ 0.38604,0.38604,0.7341,-0.5894,
+ PROTCNT=3,2,
+ RESNAME='System: Unknown','Residue: LYS 2',
+ RESSTATE=0,
+ STATEINF(0)%FIRST_ATOM=7, STATEINF(0)%FIRST_CHARGE=0, 
+ STATEINF(0)%FIRST_STATE=0, STATEINF(0)%NUM_ATOMS=22, STATEINF(0)%NUM_STATES=2, 
+ STATENE=-0.845507374619,0,
+ TRESCNT=1,CPHFIRST_SOL=35, CPH_IGB=2, CPH_INTDIEL=1.0, 
+/
