+++ constph/calibration-systems/lys-implicit.state.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".19894260496960556" y=".12235987845896074" z=".07083757598853216"/>
+		<Position x=".20425252744076314" y=".2169063944369947" z=".01686288672143928"/>
+		<Position x=".12964440868553892" y=".28582046969823743" z=".05642261334614707"/>
+		<Position x=".18524395766006552" y=".19964460045352792" z="-.08906727504605826"/>
+		<Position x=".34184765996223865" y=".27734358608886384" z=".03358591911942005"/>
+		<Position x=".4271195475235977" y=".21871808198324424" z=".10043464357205409"/>
+		<Position x=".36281425687326796" y=".39445673275741794" z="-.02715772276438853"/>
+		<Position x=".2853218084852187" y=".43375969051885016" z="-.0786583345325704"/>
+		<Position x=".4885732345169574" y=".4715260145811073" z="-.02457786865052569"/>
+		<Position x=".5445442388926345" y=".444182623621003" z=".06487448978219117"/>
+		<Position x=".5723258961187541" y=".43281059924011134" z="-.1480490245764764"/>
+		<Position x=".5813957797017808" y=".3242243917118859" z="-.15100956465248142"/>
+		<Position x=".5204799910480834" y=".4650278645630681" z="-.23835983884252016"/>
+		<Position x=".7137504267723773" y=".49277656588061874" z="-.14689915530199915"/>
+		<Position x=".707458398865405" y=".6014532556230128" z="-.15247959945344053"/>
+		<Position x=".7639057801697277" y=".46468652512366276" z="-.05428816624178779"/>
+		<Position x=".7958691975397569" y=".4420057070155808" z="-.26605926369370736"/>
+		<Position x=".803795521038301" y=".3334172733988673" z="-.2607860659655204"/>
+		<Position x=".7457064732980284" y=".46972356301059864" z="-.3587819888581418"/>
+		<Position x=".936146346270332" y=".5038036322273636" z="-.2631730237722632"/>
+		<Position x=".9266959191345466" y=".612363379243236" z="-.2662388398940563"/>
+		<Position x=".9843209314042226" y=".4770051677146486" z="-.16912529354709405"/>
+		<Position x="1.018748923174159" y=".45694250408541776" z="-.3775150486622296"/>
+		<Position x=".9767797554300689" y=".48205067195209184" z="-.46588999486922494"/>
+		<Position x="1.110913986675965" y=".49817542406223786" z="-.37478407042897305"/>
+		<Position x="1.0302340465242468" y=".35660658887363395" z="-.3758534736710918"/>
+		<Position x=".46149265039736065" y=".6226625521313681" z="-.01786768692289489"/>
+		<Position x=".3651952951936088" y=".6716690019554814" z="-.0777070494046201"/>
+		<Position x=".5471431004717195" y=".6964476059096201" z=".0535684758998604"/>
+		<Position x=".6231665943233965" y=".6479814585161948" z=".09909275122337623"/>
+		<Position x=".5392996595416539" y=".8416661617533769" z=".06994181466535913"/>
+		<Position x=".443786043600505" y=".869228846766415" z=".11465492310205039"/>
+		<Position x=".6198147238779415" y=".8764067767569496" z=".1346889750064995"/>
+		<Position x=".5483753978368496" y=".8907469390083587" z="-.026962889710202196"/>
+	</Positions>
+</State>
