+++ constph/calibration-systems/tyr-implicit.cpin
@@ -1,0 +1,13 @@
+&CNSTPH
+ CHRGDAT=-0.4157,0.2719,-0.0014,0.0876,-0.0152,0.0295,0.0295,-0.0011,-0.1906,
+ 0.1699,-0.2341,0.1656,0.3226,-0.5579,0.3992,-0.2341,0.1656,-0.1906,0.1699,
+ 0.5973,-0.5679,-0.4157,0.2719,-0.0014,0.0876,-0.0858,0.019,0.019,-0.213,-0.103,
+ 0.132,-0.498,0.132,0.777,-0.814,0.0,-0.498,0.132,-0.103,0.132,0.5973,-0.5679,
+ PROTCNT=1,0,
+ RESNAME='System: Unknown','Residue: TYR 2',
+ RESSTATE=0,
+ STATEINF(0)%FIRST_ATOM=7, STATEINF(0)%FIRST_CHARGE=0, 
+ STATEINF(0)%FIRST_STATE=0, STATEINF(0)%NUM_ATOMS=21, STATEINF(0)%NUM_STATES=2, 
+ STATENE=0,-78.310003685,
+ TRESCNT=1,
+/
