+++ constph/calibration-systems/tyr-implicit.state.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" ?>
+<State openmmVersion="7.0" time="0" type="State" version="1">
+	<PeriodicBoxVectors>
+		<A x="2" y="0" z="0"/>
+		<B x="0" y="2" z="0"/>
+		<C x="0" y="0" z="2"/>
+	</PeriodicBoxVectors>
+	<Positions>
+		<Position x=".1788545955942833" y=".14562957636174859" z=".1022611420341942"/>
+		<Position x=".1933265427824649" y=".2256759675531461" z=".029713240648567226"/>
+		<Position x=".11425234445501681" y=".29986622524302475" z=".040825220035739546"/>
+		<Position x=".19062426416442144" y=".18451531928455672" z="-.07117715631362442"/>
+		<Position x=".3272332118195919" y=".291738914448606" z=".05297902968548022"/>
+		<Position x=".40037034674102717" y=".2527658038137825" z=".14398104267081588"/>
+		<Position x=".35874832706815224" y=".39153943939305436" z="-.02940064404121935"/>
+		<Position x=".291374364689848" y=".416874767664347" z="-.10025210829674687"/>
+		<Position x=".48284793630932554" y=".4699618723580573" z="-.02388400940943562"/>
+		<Position x=".5202429517203131" y=".46964828708029643" z=".07850318076090244"/>
+		<Position x=".5882505567040278" y=".40327228114558533" z="-.11374627848542791"/>
+		<Position x=".59949327037916" y=".29918538138263634" z="-.08339860126678182"/>
+		<Position x=".5521903101592743" y=".40350918460183127" z="-.21661175951984069"/>
+		<Position x=".7250265944710771" y=".4687403822502156" z="-.10843423650531493"/>
+		<Position x=".7689704366363495" y=".5500793066840491" z="-.214777844246814"/>
+		<Position x=".7059503168196208" y=".5648105363318067" z="-.30125097178866667"/>
+		<Position x=".894755110031719" y=".6131015365025428" z="-.2083339208363878"/>
+		<Position x=".9281894884197277" y=".6755449551526133" z="-.2898680352650705"/>
+		<Position x=".9770775387247372" y=".5947856295598573" z="-.09506371912446794"/>
+		<Position x="1.0986673624435108" y=".6553181910302288" z="-.08802067205070986"/>
+		<Position x="1.1167078911526174" y=".707559892084294" z="-.1665465297935304"/>
+		<Position x=".9328659238935662" y=".5133501529003307" z=".011392127730935848"/>
+		<Position x=".9962023633235376" y=".4997205607361001" z=".09780636790113936"/>
+		<Position x=".8070443658474286" y=".4504944462036156" z=".004603169551552925"/>
+		<Position x=".7734126147300601" y=".38828788816669074" z=".0862458819306494"/>
+		<Position x=".4597770953874184" y=".6169538689848062" z="-.0643983603759361"/>
+		<Position x=".36811145845456084" y=".6478799228504176" z="-.14143215197199227"/>
+		<Position x=".5437049464725646" y=".7080194016786284" z="-.013362320889766187"/>
+		<Position x=".6178960422728494" y=".6745847870352802" z=".046468289130368046"/>
+		<Position x=".5391309499667037" y=".8512870628279583" z="-.042159319492910886"/>
+		<Position x=".4420464257692627" y=".8920915445397612" z="-.014025758695341587"/>
+		<Position x=".6168597545820496" y=".9031662886149686" z=".013954942052828885"/>
+		<Position x=".5551791701315895" y=".8683995802679596" z="-.14860909224689525"/>
+	</Positions>
+</State>
