+++ constph/tests/test_implicit.py
@@ -1,7 +1,8 @@
 from __future__ import print_function
 from simtk import unit, openmm
 from simtk.openmm import app
-from constph.constph import MonteCarloTitration, CalibrationTitration, MBarCalibrationTitration
+from constph.constph import MonteCarloTitration
+from constph.calibration import CalibrationTitration, MBarCalibrationTitration
 from unittest import TestCase, skip
 from . import get_data
