+++ setup.py
@@ -8,9 +8,6 @@ DOCLINES = __doc__.split("\n")
 
 import os
 import sys
-import glob
-import traceback
-import numpy as np
 from os.path import join as pjoin
 from os.path import relpath
 from setuptools import setup, Extension, find_packages
@@ -28,6 +25,7 @@ if '--debug' in sys.argv:
     DEBUG = True
 else:
     DEBUG = False
+
 
 def find_package_data(data_root, package_root):
     files = []
@@ -60,18 +58,35 @@ Programming Language :: Python :: 3.5
 
 extensions = []
 
-setup(name='constph',
-      author='Bas Rustenburg',
-      author_email='bas.rustenburg@choderalab.org',
-      description=DOCLINES[0],
-      long_description="\n".join(DOCLINES[2:]),
-      version=__version__,
-      url='https://github.com/choderalab/openmm-constph',
-      platforms=['Linux', 'Mac OS-X', 'Unix'],
-      classifiers=CLASSIFIERS.splitlines(),
-      packages=['constph', 'constph.tests'],
-      package_data={'constph' : find_package_data('constph/examples', 'constph') + find_package_data('constph/tests/testsystems', 'constph')}, 
-      zip_safe=False,
-      ext_modules=extensions,
-      test_suite='nose.collector',
-      )
+setup(
+    name='constph',
+    author='Bas Rustenburg',
+    author_email='bas.rustenburg@choderalab.org',
+    description=DOCLINES[0],
+    long_description="\n".join(
+        DOCLINES[
+            2:]),
+    version=__version__,
+    url='https://github.com/choderalab/openmm-constph',
+    platforms=[
+        'Linux',
+        'Mac OS-X',
+        'Unix'],
+    classifiers=CLASSIFIERS.splitlines(),
+    packages=[
+        'constph',
+        'constph.tests'],
+    package_data={
+        'constph': find_package_data(
+            'constph/examples',
+            'constph') +
+        find_package_data(
+            'constph/tests/testsystems',
+            'constph') +
+        find_package_data(
+            'constph/calibration-systems',
+            'constph')} ,
+    zip_safe=False,
+    ext_modules=extensions,
+    test_suite='nose.collector',
+)
